requirements.txt
discord.py==2.3.2
py-cord==2.4.0
aiohttp==3.9.1
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
cryptography==41.0.8
asyncpg==0.29.0
pydantic==2.5.2
python-multipart==0.0.6
fastapi==0.104.1
uvicorn==0.24.0
